working on Demo code

+++ TG68_AMR_Software/Apps/Demo/main.c
@@ -322,7 +322,7 @@ int main(int argc,char *argv)
 
 	FrameBuffer=(short *)malloc(sizeof(short)*640*960+15);
 	FrameBuffer=(short *)(((int)FrameBuffer+15)&~15); // Align to nearest 16 byte boundary.
-	HW_VGA_L(FRAMEBUFFERPTR)=FrameBuffer;
+	HW_VGA_L(FRAMEBUFFERPTR) = (long unsigned int) FrameBuffer;
 
 	memset(FrameBuffer,0,sizeof(short)*640*960);
